+++ frontend/src/components/tasks/AddTask.vue
@@ -70,7 +70,7 @@ import QuickAddMagic from '@/components/tasks/partials/QuickAddMagic.vue'
 import {parseSubtasksViaIndention} from '@/helpers/parseSubtasksViaIndention'
 import TaskRelationService from '@/services/taskRelation'
 import TaskRelationModel from '@/models/taskRelation'
-import {getLabelsFromPrefix} from '@/modules/parseTaskText'
+import {getLabelsFromPrefix} from '@/modules/quickAddMagic'
 
 import {useAuthStore} from '@/stores/auth'
 import {useTaskStore} from '@/stores/tasks'
