+++ frontend/src/modelTypes/IUserSettings.ts
@@ -1,6 +1,6 @@
 import type {IAbstract} from './IAbstract'
 import type {IProject} from './IProject'
-import type {PrefixMode} from '@/modules/parseTaskText'
+import type {PrefixMode} from '@/modules/quickAddMagic'
 import type {BasicColorSchema} from '@vueuse/core'
 import type {SupportedLocale} from '@/i18n'
 import type {DefaultProjectViewKind} from '@/modelTypes/IProjectView'
