+++ frontend/src/modules/quickAddMagic/index.ts
@@ -1,4 +1,4 @@
-export {parseTaskText} from './parseTaskText'
+export {parseTaskText} from './quickAddMagic'
 export {PrefixMode, PREFIXES} from './prefixes'
 export {getLabelsFromPrefix, getProjectFromPrefix} from './prefixParser'
 export {cleanupItemText} from './textCleanup'
