+++ frontend/src/modules/quickAddMagic/quickAddMagic.test.ts
