+++ frontend/src/modules/quickAddMagic/quickAddMagic.ts
