+++ frontend/src/modules/quickAddMagic/repeatParser.ts
