+++ frontend/src/modules/quickAddMagic/textCleanup.ts
