+++ frontend/src/modules/quickAddMagic/types.ts
